Изменение структуры папок фронтенд-адаптера
